* new ObjectJcpp_Test
* Tests with DEF_ObjectJc_SIMPLE, type check possible with DEF_ObjectJc_REFLREF
* Eclipse-GCC-Project
* new src/test/ZmakeGcc
*  All testAll_ObjectJcpp_emCBase well tested
* new src_emC/, mklib
* Test works with gradle

+++ src/test/VS15/StateM_Test/+clean.bat
@@ -1,11 +1,9 @@
-rmdir /S/Q debug
-rmdir /S/Q x64
-rmdir /S/Q dbg_c1
-rmdir /S/Q x32
-rmdir /S/Q .vs
-del /S/Q *.user
-del /S/Q *.db
-del /S/Q *.sdf
-
-pause
+if exist debug rmdir /S/Q debug
+if exist dbgC1 rmdir /S/Q dbgC1
+if exist x64 rmdir /S/Q x64
+if exist x32 rmdir /S/Q x32
+if exist .vs rmdir /S/Q .vs
+if exist *.user del /S/Q *.user
+if exist *.db del /S/Q *.db
+if exist *.sdf del /S/Q *.sdf
